+++ tests/test_bresenham.cpp
@@ -1,9 +1,8 @@
 
-#include <libtcod/bresenham.h>
-
 #include <algorithm>
 #include <array>
 #include <catch2/catch_all.hpp>
+#include <libtcod/bresenham.hpp>
 #include <vector>
 
 /**