+++ tests/test_color.cpp
@@ -1,6 +1,4 @@
 
-#include <libtcod/color.h>
-
 #include <catch2/catch_all.hpp>
 #include <libtcod/color.hpp>
 #include <libtcod/console.hpp>